refactor(cli): extract sensor findings owner
